requirement.txt
# Microsoft Agent Framework (pinned to avoid breaking changes)
agent-framework-core==1.0.0b260107
agent-framework-azure-ai==1.0.0b260107

# Azure services
azure-storage-blob>=12.23.0
azure-identity>=1.19.0

# Utilities
python-dotenv>=1.0.1
requests>=2.32.0

# ── Grounding Pipeline Dependencies ──
# (Open Food Facts, USDA, FDA, DDInter, FooDB, DSLB APIs all use 'requests' — no extra deps)